Feature/dashboard module

+++ apps/api/src/auth/auth.service.ts
@@ -1,10 +1,12 @@
 import {
+  BadRequestException,
   ForbiddenException,
   Injectable,
   NotFoundException,
   UnauthorizedException,
 } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
+import { Role } from '@prisma/client';
 import * as bcrypt from 'bcrypt';
 import { TokensService } from 'src/tokens/tokens.service';
 import { UsersService } from 'src/users/users.service';
@@ -20,21 +22,21 @@ export class AuthService {
     private readonly tokensService: TokensService,
   ) {}
 
-  async generateAccessToken(userId: number, email: string) {
-    const payload = { sub: userId, email };
+  async generateAccessToken(userId: number, email: string, role: Role) {
+    const payload = { sub: userId, email, role };
 
-    return this.jwtService.signAsync(payload, {
+    return this.jwtService.sign(payload, {
       expiresIn: 60 * 15 * 1000,
       secret: jwtSecret,
     });
   }
 
-  async generateRefreshToken(userId: number, email: string) {
-    const payload = { sub: userId, email };
+  async generateRefreshToken(userId: number, email: string, role: Role) {
+    const payload = { sub: userId, email, role };
 
-    return this.jwtService.signAsync(payload, {
-      expiresIn: 60 * 60 * 24 * 1000,
-      secret: jwtSecret,
+    return this.jwtService.sign(payload, {
+      expiresIn: '1d',
+      secret: jwtRefreshSecret,
     });
   }
 
@@ -55,8 +57,16 @@ export class AuthService {
       );
     }
 
-    const accessToken = await this.generateAccessToken(user.id, user.email);
-    const refreshToken = await this.generateRefreshToken(user.id, user.email);
+    const accessToken = await this.generateAccessToken(
+      user.id,
+      user.email,
+      user.role,
+    );
+    const refreshToken = await this.generateRefreshToken(
+      user.id,
+      user.email,
+      user.role,
+    );
 
     await this.tokensService.save(user.id, refreshToken);
 
@@ -68,7 +78,26 @@ export class AuthService {
     };
   }
 
+  async register(input: {
+    name: string;
+    email: string;
+    password: string;
+    role: Role;
+    surname: string;
+  }) {
+    const isUserExists = await this.userService.findByEmail(input.email);
+
+    if (isUserExists) {
+      throw new BadRequestException('User already exists');
+    }
+    return this.userService.create(input);
+  }
+
   async logout(userId: number) {
+    const isTokenExist = await this.tokensService.findOne(userId);
+    if (!isTokenExist) {
+      throw new UnauthorizedException('Access denied');
+    }
     return this.tokensService.remove(userId);
   }
 
@@ -80,14 +109,26 @@ export class AuthService {
       throw new ForbiddenException('Access denied');
     }
 
-    const isRfTokensMatches = bcrypt.compare(refreshToken, refreshToken);
+    const isRfTokensMatches = bcrypt.compare(
+      rfToken.refreshToken,
+      refreshToken,
+    );
 
     if (!isRfTokensMatches) {
       throw new ForbiddenException('Access denied');
     }
 
-    const newRfToken = await this.generateRefreshToken(user.id, user.email);
-    const newAccessToken = await this.generateAccessToken(user.id, user.email);
+    const newRfToken = await this.generateRefreshToken(
+      user.id,
+      user.email,
+      user.role,
+    );
+
+    const newAccessToken = await this.generateAccessToken(
+      user.id,
+      user.email,
+      user.role,
+    );
 
     return {
       accessToken: newAccessToken,
